move logic at node startup
